Move validators and OAuth outside of client
