+++ composer.json
@@ -16,13 +16,14 @@
         }
     ],
     "require": {
-        "php": ">=5.6",
+        "php": "~7.0",
         "guzzle/guzzle": "~3.6",
+        "guzzlehttp/guzzle": "~6.0",
         "zendframework/zend-diactoros": "~1.1",
         "container-interop/container-interop": "^1.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.4"
+        "phpunit/phpunit": "~5.2"
     },
     "autoload": {
         "psr-4": {
@@ -33,5 +34,8 @@
         "psr-4": {
             "ZfrShopifyTest\\": "tests/"
         }
+    },
+    "scripts": {
+        "test": "phpunit"
     }
 }
