+++ src/Container/ShopifyClientFactory.php
@@ -32,7 +32,6 @@ use ZfrShopify\ShopifyClient;
  *     'zfr_shopify' => [
  *         'shop'          => '', // a shop name, WITHOUT the ".myshopify.com" part (only required for private apps)
  *         'api_key'       => '', // an API key, always required
- *         'shared_secret' => '', // your shared secret, used to validate any request (only required for public apps)
  *         'access_token'  => '', // an access token that you got from the OAuth dance (only required for public apps)
  *         'password'      => '', // a password retrieved from private app (only required for private apps)
  *         'private_app'   => true, // true for private app, false for apps on App Store
@@ -58,8 +57,8 @@ class ShopifyClientFactory
                 throw new RuntimeException('Options "shop", "api_key" and "password" are mandatory when creating Shopify client for a private app');
             }
         } else {
-            if (!isset($config['api_key']) || !isset($config['shared_secret'])) {
-                throw new RuntimeException('Options "api_key" and "shared_secret" are mandatory when creating Shopify client for a public app');
+            if (!isset($config['api_key'])) {
+                throw new RuntimeException('Options "api_key" is mandatory when creating Shopify client for a public app');
             }
         }
