+++ src/ShopifyClient.php
@@ -121,10 +121,6 @@ use ZfrShopify\Exception;
  * @method array createWebhook(array $args = []) {@command Shopify CreateWebhook}
  * @method array updateWebhook(array $args = []) {@command Shopify UpdateWebhook}
  * @method array deleteWebhook(array $args = []) {@command Shopify DeleteWebhook}
- *
- * OAUTH RELATED METHODS:
- *
- * @method array exchangeCodeForToken(array $args = []) {@command Shopify ExchangeCodeForToken}
  */
 class ShopifyClient extends Client
 {
@@ -170,15 +166,6 @@ class ShopifyClient extends Client
     public function setApiKey($apiKey)
     {
         $this->options['api_key'] = (string) $apiKey;
-    }
-
-    /**
-     * @param string $sharedSecret
-     * @return void
-     */
-    public function setSharedSecret($sharedSecret)
-    {
-        $this->options['shared_secret'] = (string) $sharedSecret;
     }
 
     /**
@@ -307,136 +294,7 @@ class ShopifyClient extends Client
         if ($this->options['private_app']) {
             $request->setAuth($this->options['api_key'], $this->options['password']);
         } else {
-            // There is a special case for the "ExchangeCodeForToken" where we authorize the request differently
-            if ($command->getName() !== 'ExchangeCodeForToken') {
-                $request->setHeader('X-Shopify-Access-Token', $this->options['access_token']);
-            }
+            $request->setHeader('X-Shopify-Access-Token', $this->options['access_token']);
         }
-    }
-
-    /**
-     * Validate the incoming request and check if it is valid
-     *
-     * @link   https://docs.shopify.com/api/authentication/oauth#verification
-     * @param  ServerRequestInterface $request
-     * @return void
-     * @throws Exception\InvalidRequestException
-     */
-    public function validateRequest(ServerRequestInterface $request)
-    {
-        // First step: extract the query params
-        $queryParams = $request->getQueryParams();
-
-        $this->validateShopHostname($queryParams);
-        $this->validateHmac($queryParams);
-    }
-
-    /**
-     * Validate the webhook coming from Shopify
-     *
-     * @link   https://docs.shopify.com/api/webhooks/using-webhooks#verify-webhook
-     * @param  ServerRequestInterface $request
-     * @return void
-     * @throws Exception\InvalidWebhookException
-     */
-    public function validateWebhook(ServerRequestInterface $request)
-    {
-        $hmac = $request->getHeaderLine('X-Shopify-Hmac-SHA256');
-
-        if (empty($hmac)) {
-            throw new Exception\InvalidRequestException('Incoming Shopify webhook could not be validated');
-        }
-
-        $data           = new PhpInputStream();
-        $calculatedHmac = base64_encode(hash_hmac('sha256', $data->getContents(), $this->options['shared_secret'], true));
-
-        if (hash_equals($hmac, $calculatedHmac)) {
-            return;
-        }
-
-        throw new Exception\InvalidRequestException('Incoming Shopify webhook could not be validated');
-    }
-
-    /**
-     * Create an authorization redirection request
-     *
-     * Please note that this method will automatically generate a nonce value. You are responsible to
-     * persist it in database, and validate it during the OAuth dance
-     *
-     * @param  string $shopDomain
-     * @param  array  $scopes
-     * @param  string $redirectionUri
-     * @param  string $nonce
-     * @return ResponseInterface
-     * @throws Exception\MissingApiKeyException
-     */
-    public function createAuthorizationResponse($shopDomain, array $scopes, $redirectionUri, $nonce)
-    {
-        $uri = sprintf(
-            'https://%s.myshopify.com/admin/oauth/authorize?client_id=%s&scope=%s&redirect_uri=%s&state=%s',
-            str_replace('.myshopify.com', '', $shopDomain),
-            $this->options['api_key'],
-            implode(',', $scopes),
-            $redirectionUri,
-            $nonce
-        );
-
-        return new RedirectResponse($uri);
-    }
-
-    /**
-     * According to Shopify, a shop hostname must ends by "myshopify.com", and must only contains
-     * letters, numbers, dots and hyphens
-     *
-     * @param  array $queryParams
-     * @return void
-     * @throws Exception\InvalidRequestException
-     */
-    private function validateShopHostname(array $queryParams)
-    {
-        $shop = isset($queryParams['shop']) ? $queryParams['shop'] : '';
-
-        if (preg_match('/^[a-zA-Z0-9.-]*(myshopify.com)$/', $shop) === 1) {
-            return;
-        }
-
-        throw new Exception\InvalidRequestException('Incoming request from Shopify could not be validated');
-    }
-
-    /**
-     * Validate the given HMAC
-     *
-     * @param  array  $queryParams
-     * @return void
-     * @throws Exception\InvalidRequestException
-     */
-    private function validateHmac(array $queryParams)
-    {
-        $expectedHmac = isset($queryParams['hmac']) ? $queryParams['hmac'] : '';
-
-        // First step: remove HMAC and signature keys
-        unset($queryParams['hmac'], $queryParams['signature']);
-
-        // Second step: keys are sorted lexicographically
-        ksort($queryParams);
-
-        $pairs = [];
-
-        foreach ($queryParams as $key => $value) {
-            // Third step: "&" and "%" are replaced by "%26" and "%25" in keys and values, and in addition
-            // "=" is replaced by "%3D" in keys
-            $key   = strtr($key, ['&' => '%26', '%' => '%25', '=' => '%3D']);
-            $value = strtr($value, ['&' => '%26', '%' => '%25']);
-
-            $pairs[] = $key . '=' . $value;
-        }
-
-        $key = implode('&', $pairs);
-
-        if (hash_equals($expectedHmac, hash_hmac('sha256', $key, $this->options['shared_secret']))) {
-            return;
-        };
-
-        throw new Exception\InvalidRequestException('Incoming request from Shopify could not be validated');
     }
 }
